Allow matching prefixes of 4 characters
BIP-39 specifies that the wordlist should only contains words that
are different on their first 4 characters, for all words. This allows
to only check the prefixes if the length of a word is 4. Otherwise
the behaviour is unchanged.

This does not seem to impact the performance significantly (I noticed
less than 1% change in runtime of `npm run unit`).